server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
httpx
python-dotenv
pydantic
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
greenlet
aiosmtplib
aiogram>=3.13,<4.0
